fix:링크 수정 및 pep8적용
