+++ spartamarket/products/urls.py
@@ -7,5 +7,5 @@ urlpatterns = [
     path('<int:pk>/', views.details, name='details'),
     path('<int:pk>/delete/', views.delete, name='delete'),
     path('<int:pk>/update/', views.update, name='update'),
-
+    path('<int:pk>/comments/', views.comments, name='comments'),
 ]
