+++ spartamarket/users/views.py
@@ -1,12 +1,5 @@
-from django.shortcuts import render, redirect
-from django.views.decorators.http import require_POST, require_http_methods
-from django.contrib.auth import logout as auth_logout
-
-
-
-# @require_http_methods(["GET","POST"]) 
-# def users(request):
-#     return render(request, "users/users.html")
+from django.shortcuts import render
+from django.views.decorators.http import require_http_methods
 
 
 @require_http_methods(["GET","POST"]) 
@@ -15,9 +8,6 @@ def profile(request, username):
         "username": username,
     }
     return render(request, "users/profile.html", context)
-
-def good(request):
-    pass
 
 
 def product(request, username):
